pyproject.toml
[project]
name = "ai-influencer"
version = "0.1.0"
description = "Automated video generation pipeline with fail-loud safety guards"
requires-python = ">=3.11"
readme = "README.md"
license = { text = "MIT" }

dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "httpx>=0.25.0",
    "tenacity>=8.2.0",
    "python-dotenv>=1.0.0",
    "pillow>=10.1.0",
    "opencv-python>=4.8.0",
    "apscheduler>=3.10.0",
    "slowapi>=0.1.9",
    "google-cloud-aiplatform>=1.122.0",
    "google-auth>=2.41.1",
    "pytz>=2025.2",
]

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "hypothesis>=6.92.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend/app"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = []

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
disallow_untyped_decorators = false
warn_redundant_casts = true
warn_unused_ignores = true
